Move ReasonCode to MarketActivityRecord in ingretion
